requirements.txt
pyrogram==2.0.106
deep_translator
python-dotenv